+++ next/src/components/note/MarkdownText.tsx
@@ -1,71 +1,20 @@
 import Image from 'next/image'
-import React from 'react'
 import ReactMarkdown from 'react-markdown'
 import type { Components } from 'react-markdown'
-import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter'
-import bash from 'react-syntax-highlighter/dist/cjs/languages/prism/bash'
-import python from 'react-syntax-highlighter/dist/cjs/languages/prism/python'
-import ruby from 'react-syntax-highlighter/dist/cjs/languages/prism/ruby'
-import typescript from 'react-syntax-highlighter/dist/cjs/languages/prism/typescript'
-import yaml from 'react-syntax-highlighter/dist/cjs/languages/prism/yaml'
-import { oneDark } from 'react-syntax-highlighter/dist/cjs/styles/prism'
 import rehypeRaw from 'rehype-raw'
 import rehypeSanitize from 'rehype-sanitize'
 import remarkGfm from 'remark-gfm'
+import CodeBlock from './CodeBlock'
 import styles from '@/styles/MarkdownText.module.css'
-
-SyntaxHighlighter.registerLanguage('ruby', ruby)
-SyntaxHighlighter.registerLanguage('yaml', yaml)
-SyntaxHighlighter.registerLanguage('bash', bash)
-SyntaxHighlighter.registerLanguage('typescript', typescript)
-SyntaxHighlighter.registerLanguage('python', python)
 
 interface MarkdownTextProps {
   content: string
   className?: string
 }
 
-interface CustomCodeProps {
-  className?: string
-  children?: React.ReactNode
-}
-
 const MarkdownText = ({ content, className = '' }: MarkdownTextProps) => {
   const customComponents: Components = {
-    code(props: CustomCodeProps) {
-      const { className, children } = props
-      const isInline = !String(children).includes('\n')
-      if (isInline) {
-        return <code className={className}>{children}</code>
-      }
-
-      const match = /language-(\w+)/.exec(className || '')
-      const language = match ? match[1] : 'plaintext'
-
-      const languageMap: { [key: string]: string } = {
-        rb: 'ruby',
-        ruby: 'ruby',
-        yml: 'yaml',
-        yaml: 'yaml',
-        bash: 'bash',
-        sh: 'bash',
-        shell: 'bash',
-        zsh: 'bash',
-        ts: 'typescript',
-        typescript: 'typescript',
-        py: 'python',
-        python: 'python',
-        pyc: 'python',
-      }
-
-      const mappedLanguage = languageMap[language.toLowerCase()] || language
-
-      return (
-        <SyntaxHighlighter style={oneDark} language={mappedLanguage}>
-          {String(children).replace(/\n$/, '')}
-        </SyntaxHighlighter>
-      )
-    },
+    code: CodeBlock,
     img({ src, alt = 'image', width }) {
       if (!src) return null
       if (typeof width !== 'number') return null
